+++ elijah-core/src/test/java/tripleo/elijah/SimpleTest.java
@@ -1,6 +1,7 @@
 package tripleo.elijah;
 
 import tripleo.elijah.comp.Compilation0;
+import tripleo.elijah.test_help.AssertingLiveClass;
 
 public interface SimpleTest {
 	SimpleTest setFile(String aS);
