+++ elijah-core/src/test/java/tripleo/elijah/test_help/AssertingLiveClass.java
@@ -1,4 +1,4 @@
-package tripleo.elijah;
+package tripleo.elijah.test_help;
 
 public class AssertingLiveClass {
 	private String  className;
